Add cluster with manager
